+++ src/commands/projects/ProjectCommand.ts
@@ -3,6 +3,8 @@ import CrossingGuardBot from "../../CrossingGuardBot";
 import { ProjectStatus } from "../../ProjectStatus";
 import ProjectAttachment from "../../ProjectAttachment";
 import ProjectLink from "../../ProjectLink";
+import ProjectStaff from "../../ProjectStaff";
+import { ProjectStaffRank } from "../../ProjectStaffRank";
 
 const data = new SlashCommandBuilder()
     .setName("project")
@@ -13,7 +15,7 @@ const data = new SlashCommandBuilder()
         subcommand.setName("create")
             .setDescription("Create a new project with the given name")
             .addStringOption(option =>
-                option.setName("name")
+                option.setName("project")
                     .setDescription("The internal name for this project")
                     .setRequired(true))
             .addStringOption(option =>
@@ -25,7 +27,7 @@ const data = new SlashCommandBuilder()
         subcommand.setName("addexisting")
             .setDescription("Add an existing project with the given name")
             .addStringOption(option =>
-                option.setName("name")
+                option.setName("project")
                     .setDescription("The internal name for this project")
                     .setRequired(true))
             .addStringOption(option =>
@@ -49,7 +51,7 @@ const data = new SlashCommandBuilder()
                 subcommand.setName("add")
                     .setDescription("Add a link to a project")
                     .addStringOption(option =>
-                        option.setName("project_name")
+                        option.setName("project")
                             .setDescription("The name of the project")
                             .setRequired(true))
                     .addStringOption(option =>
@@ -59,8 +61,69 @@ const data = new SlashCommandBuilder()
                     .addStringOption(option =>
                         option.setName("url")
                             .setDescription("The URL for this link")
+                            .setRequired(true)))
+            .addSubcommand(subcommand =>
+                subcommand.setName("remove")
+                    .setDescription("Remove a link from a project")
+                    .addStringOption(option =>
+                        option.setName("project")
+                            .setDescription("The name of the project")
+                            .setRequired(true))
+                    .addStringOption(option =>
+                        option.setName("name")
+                            .setDescription("The name to display for this link")
+                            .setRequired(true)))
+            // List Links Subcommand
+            .addSubcommand(subcommand =>
+                subcommand.setName("list")
+                    .setDescription("View a list of a project's links")
+                    .addStringOption(option =>
+                        option.setName("project")
+                            .setDescription("The name of the project")
                             .setRequired(true))))
     // Set Field Subcommand Group
+    .addSubcommandGroup(subcommandGroup =>
+        subcommandGroup.setName("staff")
+            .setDescription("Manage this project's staff")
+            // Add Staff Subcommand
+            .addSubcommand(subcommand =>
+                subcommand.setName("add")
+                    .setDescription("Add a staff member to a project")
+                    .addStringOption(option =>
+                        option.setName("project")
+                            .setDescription("The name of the project")
+                            .setRequired(true))
+                    .addUserOption(option =>
+                        option.setName("user")
+                            .setDescription("The user to make staff")
+                            .setRequired(true))
+                    .addStringOption(option =>
+                        option.setName("rank")
+                            .setDescription("The rank of this staff member")
+                            .setRequired(true)
+                            .addChoices(
+                                { name: "Lead", value: "0" },
+                                { name: "Staff", value: "1" },
+                            )))
+            .addSubcommand(subcommand =>
+                subcommand.setName("remove")
+                    .setDescription("Remove a staff member from a project")
+                    .addStringOption(option =>
+                        option.setName("project")
+                            .setDescription("The name of the project")
+                            .setRequired(true))
+                    .addUserOption(option =>
+                        option.setName("user")
+                            .setDescription("The staff user to remove from this project")
+                            .setRequired(true)))
+            // List Staff Subcommand
+            .addSubcommand(subcommand =>
+                subcommand.setName("list")
+                    .setDescription("View a list of a project's staff")
+                    .addStringOption(option =>
+                        option.setName("project")
+                            .setDescription("The name of the project")
+                            .setRequired(true))))
     .addSubcommandGroup(subcommandGroup =>
         subcommandGroup.setName("set")
             .setDescription("Set one of this project's fields")
@@ -69,7 +132,7 @@ const data = new SlashCommandBuilder()
                 subcommand.setName("ip")
                     .setDescription("Set a project's ip & version")
                     .addStringOption(option =>
-                        option.setName("name")
+                        option.setName("project")
                             .setDescription("The name of the project")
                             .setRequired(true))
                     .addStringOption(option =>
@@ -81,7 +144,7 @@ const data = new SlashCommandBuilder()
                 subcommand.setName("status")
                     .setDescription("Set a project's status")
                     .addStringOption(option =>
-                        option.setName("name")
+                        option.setName("project")
                             .setDescription("The name of the project")
                             .setRequired(true))
                     .addStringOption(option =>
@@ -99,7 +162,7 @@ const data = new SlashCommandBuilder()
                 subcommand.setName("emoji")
                     .setDescription("Set a project's emoji")
                     .addStringOption(option =>
-                        option.setName("name")
+                        option.setName("project")
                             .setDescription("The name of the project")
                             .setRequired(true))
                     .addStringOption(option =>
@@ -111,7 +174,7 @@ const data = new SlashCommandBuilder()
                 subcommand.setName("description")
                     .setDescription("Set a project's description")
                     .addStringOption(option =>
-                        option.setName("name")
+                        option.setName("project")
                             .setDescription("The name of the project")
                             .setRequired(true))
                     .addStringOption(option =>
@@ -123,7 +186,7 @@ const data = new SlashCommandBuilder()
                 subcommand.setName("attachments")
                     .setDescription("Set the attachments of a project")
                     .addStringOption(option =>
-                        option.setName("name")
+                        option.setName("project")
                             .setDescription("The name of the project")
                             .setRequired(true))
                     .addStringOption(option =>
@@ -135,7 +198,7 @@ const data = new SlashCommandBuilder()
                 subcommand.setName("discord_id")
                     .setDescription("Set this project's linked disord guild ID")
                     .addStringOption(option =>
-                        option.setName("name")
+                        option.setName("project")
                             .setDescription("The name of the project")
                             .setRequired(true))
                     .addStringOption(option =>
@@ -165,6 +228,19 @@ async function execute(interaction: ChatInputCommandInteraction) {
     else if (subcommandGroup == "link") {
         if (subcommand == "add")
             executeAddLink(interaction);
+        else if (subcommand == "list")
+            executeListLinks(interaction);
+        else if (subcommand == "remove")
+            executeRemoveLink(interaction);
+    }
+
+    else if (subcommandGroup == "staff") {
+        if (subcommand == "add")
+            executeAddStaff(interaction);
+        else if (subcommand == "list")
+            executeListStaff(interaction);
+        else if (subcommand == "remove")
+            executeRemoveStaff(interaction);
     }
 
     else if (subcommand == "create")
@@ -173,8 +249,119 @@ async function execute(interaction: ChatInputCommandInteraction) {
         executeAddExistingProject(interaction);
 }
 
+async function executeRemoveStaff(interaction: ChatInputCommandInteraction) {
+    const projectName = interaction.options.getString("project");
+    const user = interaction.options.getUser("user");
+
+    if (!projectName || !user)
+        return;
+
+    CrossingGuardBot.getInstance().database.getProjectByName(projectName).then(project => {
+        if (!project) {
+            interaction.reply({ content: `No project matched the name ${projectName}`, ephemeral: true });
+            return;
+        }
+
+        project.staff = project.staff.filter(staff => staff.discordUserId !== user.id);
+        CrossingGuardBot.getInstance().database.saveProject(project);
+        CrossingGuardBot.getInstance().database.updateStaffRoles(user.id);
+
+        interaction.reply({ content: `Removed the user ${user} from ${project.displayName}`, allowedMentions: { parse: [] }, ephemeral: true });
+    });
+}
+
+async function executeRemoveLink(interaction: ChatInputCommandInteraction) {
+    const projectName = interaction.options.getString("project");
+    const linkName = interaction.options.getString("name");
+
+    if (!projectName || !linkName)
+        return;
+
+    CrossingGuardBot.getInstance().database.getProjectByName(projectName).then(project => {
+        if (!project) {
+            interaction.reply({ content: `No project matched the name ${projectName}`, ephemeral: true });
+            return;
+        }
+
+        project.links = project.links.filter(link => link.linkName !== linkName);
+        CrossingGuardBot.getInstance().database.saveProject(project);
+        interaction.reply({ content: `Removed the link \`${linkName}\` from ${project.displayName}`, ephemeral: true });
+    });
+
+}
+
+async function executeListStaff(interaction: ChatInputCommandInteraction) {
+    const projectName = interaction.options.getString("project");
+
+    if (!projectName)
+        return;
+
+    CrossingGuardBot.getInstance().database.getProjectByName(projectName).then(project => {
+        if (!project) {
+            interaction.reply({ content: `No project matched the name ${projectName}`, ephemeral: true });
+            return;
+        }
+
+        let reply = project.displayName + "'s Staff\n--------------------\n";
+
+        project.staff.forEach(staff => {
+            reply += `- <@${staff.discordUserId}> ~ ${ProjectStaffRank[staff.rank]}\n`;
+        });
+
+        interaction.reply({ content: reply, allowedMentions: { parse: [] } });
+    });
+}
+
+async function executeAddStaff(interaction: ChatInputCommandInteraction) {
+    const projectName = interaction.options.getString("project");
+    const user = interaction.options.getUser("user");
+    const rank = interaction.options.getString("rank");
+
+    if (!projectName || !user || !rank)
+        return;
+
+    CrossingGuardBot.getInstance().database.getProjectByName(projectName).then(project => {
+        if (!project) {
+            interaction.reply({ content: `No project matched the name ${projectName}`, ephemeral: true });
+            return;
+        }
+
+
+        if (project.addStaff(new ProjectStaff(project.id, user.id, +rank))) {
+            CrossingGuardBot.getInstance().database.saveProject(project);
+            CrossingGuardBot.getInstance().database.updateStaffRoles(user.id);
+
+            interaction.reply({ content: `Added ${user.toString()} to the staff of ${project.displayName} as a ${ProjectStaffRank[+rank]}`, allowedMentions: { parse: [] }, ephemeral: true });
+        } else {
+            interaction.reply({ content: `${user.toString()} is already a staff member of ${project.displayName} with the role ${ProjectStaffRank[+rank]}`, allowedMentions: { parse: [] }, ephemeral: true });
+        }
+    });
+}
+
+async function executeListLinks(interaction: ChatInputCommandInteraction) {
+    const projectName = interaction.options.getString("project");
+
+    if (!projectName)
+        return;
+
+    CrossingGuardBot.getInstance().database.getProjectByName(projectName).then(project => {
+        if (!project) {
+            interaction.reply({ content: `No project matched the name ${projectName}`, ephemeral: true });
+            return;
+        }
+
+        let reply = project.displayName + "'s Links\n--------------------\n";
+
+        project.links.forEach(link => {
+            reply += `- [${link.linkName}](${link.linkUrl})\n`;
+        });
+
+        interaction.reply(reply);
+    });
+}
+
 async function executeAddLink(interaction: ChatInputCommandInteraction) {
-    const projectName = interaction.options.getString("project_name");
+    const projectName = interaction.options.getString("project");
     const linkName = interaction.options.getString("name");
     const linkURL = interaction.options.getString("url");
 
@@ -198,7 +385,7 @@ async function executeAddLink(interaction: ChatInputCommandInteraction) {
 
 
 async function executeSetGuildID(interaction: ChatInputCommandInteraction) {
-    const projectName = interaction.options.getString("name");
+    const projectName = interaction.options.getString("project");
     const guildId = interaction.options.getString("guild_id");
 
     if (!projectName || !guildId)
@@ -218,7 +405,7 @@ async function executeSetGuildID(interaction: ChatInputCommandInteraction) {
 }
 
 async function executeSetAttachments(interaction: ChatInputCommandInteraction) {
-    const projectName = interaction.options.getString("project_name");
+    const projectName = interaction.options.getString("project");
     const msgId = interaction.options.getString("msg_with_attachments_id");
 
     if (!interaction.channel || !msgId || !projectName)
@@ -247,7 +434,7 @@ async function executeSetAttachments(interaction: ChatInputCommandInteraction) {
 }
 
 async function executeSetDescription(interaction: ChatInputCommandInteraction) {
-    const projectName = interaction.options.getString("name");
+    const projectName = interaction.options.getString("project");
     const descriptionMessageId = interaction.options.getString("description_msg_id");
 
     if (!projectName || !descriptionMessageId || !interaction.channel)
@@ -272,7 +459,7 @@ async function executeSetDescription(interaction: ChatInputCommandInteraction) {
 }
 
 async function executeSetEmoji(interaction: ChatInputCommandInteraction) {
-    const projectName = interaction.options.getString("name");
+    const projectName = interaction.options.getString("project");
     const emojiIdOrUnicode = interaction.options.getString("emoji_string");
 
     if (!projectName || !emojiIdOrUnicode)
@@ -292,7 +479,7 @@ async function executeSetEmoji(interaction: ChatInputCommandInteraction) {
 }
 
 async function executeSetStatus(interaction: ChatInputCommandInteraction) {
-    const projectName = interaction.options.getString("name");
+    const projectName = interaction.options.getString("project");
     const status = interaction.options.getString("status");
 
     if (!projectName || !status)
@@ -312,7 +499,7 @@ async function executeSetStatus(interaction: ChatInputCommandInteraction) {
 }
 
 async function executeSetIp(interaction: ChatInputCommandInteraction) {
-    const projectName = interaction.options.getString("name");
+    const projectName = interaction.options.getString("project");
     const ipString = interaction.options.getString("ip_string");
 
     if (!projectName || !ipString)
@@ -333,7 +520,7 @@ async function executeSetIp(interaction: ChatInputCommandInteraction) {
 
 
 async function executeAddExistingProject(interaction: ChatInputCommandInteraction) {
-    const projectName = interaction.options.getString("name");
+    const projectName = interaction.options.getString("project");
     const displayName = interaction.options.getString("display_name");
     const channel = interaction.options.getChannel("channel");
     const role = interaction.options.getRole("project_role");
@@ -347,7 +534,7 @@ async function executeAddExistingProject(interaction: ChatInputCommandInteractio
 }
 
 async function executeCreateProject(interaction: ChatInputCommandInteraction) {
-    const projectName = interaction.options.getString("name");
+    const projectName = interaction.options.getString("project");
     const displayName = interaction.options.getString("display_name");
 
     if (!projectName || !displayName)
